Add create_tumbleweed_vm.yml playbook
This playbook leverages the setup_libvirt_host.yml playbook to esnure
that the ALP system is ready to host Libvirt VMs, and then downloads
a Tumbleweed image and uses it to create a VM.

Updated README.md with section covering the new playbook.

Add python3-lxml to packages that need to be installed on the libvirt
host and in the Ansible container image.

Add netcat-openbsd to the packages that need to be installed on the
libvirt host.